FPHandNet/requirements.txt
onnxruntime
numpy
torch
torchvision
Pillow
